pyproject.toml
[tool.poetry]
name = "tt-cyclopedia-app"
version = "0.1.0"
description = ""
authors = ["Johany Quintero <johandquintero@utp.edu.co>"]
packages = [
    { include = "app" }
]
readme = "README.md"

[tool.poetry.scripts]
start_app = "run_app:start"
test = "pytest:main"

[tool.poetry.dependencies]
certifi = "2024.7.4"  # Explicitly pin working version
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"

sqlalchemy-orm = "^1.2.10"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0.1"
shortuuid = "^1.0.13"
pylint = "^3.2.6"
python-multipart = "^0.0.19"
gunicorn = "^23.0.0"
alembic = "^1.15.2"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
PyJWT = "^2.10.0"
pymongo = "^4.13.2"
cloudinary = "^1.44.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
alembic = "^1.15.2"
typing-extensions = "4.12.2"  # Pin this exact version

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
